up: dividido o disparo de eventos de dominio para fazer nos 2 tipos de results
