+++ src/services/Vendas/Vendas.API/Config/ApiConfig.cs
@@ -57,7 +57,8 @@ namespace Vendas.API.Config
       services.AddScoped(typeof(IPipelineBehavior<,>), typeof(LoggingBehavior<,>));
       services.AddScoped(typeof(IPipelineBehavior<,>), typeof(ValidatorBehavior<,>));
       services.AddScoped(typeof(IPipelineBehavior<,>), typeof(TransactionBehavior<,>));
-      services.AddScoped(typeof(IPipelineBehavior<,>), typeof(DispatchDomainEventsPipeline<,>));
+      services.AddScoped(typeof(IPipelineBehavior<,>), typeof(DispatchDomainEventsResultNotificationPipeline<,>));
+      services.AddScoped(typeof(IPipelineBehavior<,>), typeof(DispatchDomainEventsResultPipeline<,>));
 
       services.AddSingleton(configuration.BindSettings<EventBusSettings>(nameof(EventBusSettings)));
